% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelsummary.R
\name{accelsummary}
\alias{accelsummary}
\title{accelsummary}
\usage{
accelsummary(x, w, h = 100)
}
\arguments{
\item{x}{a vector of activity counts}

\item{w}{a vector of wear flags same length of x}

\item{h}{threshold to define active and sedenday, default h=100}
}
\value{
list of summaries
}
\description{
This function generates daily activity summaries.
}
\examples{
accelsummary(x=xi,w=wi,h=100)
}
\keyword{accelsummary}

