% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelsummary_data.R
\name{accelsummary_data}
\alias{accelsummary_data}
\title{accelsummary_data}
\usage{
accelsummary_data(data, wear, h = 100)
}
\arguments{
\item{data}{a data frame where each row is one subject day and
column 1 and 2 are id and day, and fowllowing colums
are minute level activity count}

\item{wear}{a data frame of similar strucutre to previous one containing
wear/nonwear flag}

\item{h}{threshold to define active and sedenday, default h=100}
}
\value{
data frame of subject-day level summaries
}
\description{
This function generates daily activity summaries.
}
\examples{
data(act)
accelsummary_data(data=act,wear=flag,h=100)


}
\keyword{accelsummary_data}

