% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_profile.R
\name{activity_profile}
\alias{activity_profile}
\title{activity_profile}
\usage{
activity_profile(data, wear, id, day, h = 100)
}
\arguments{
\item{data}{a data frame where each row is one subject day and
column 1 and 2 are id and day, and fowllowing colums
are minute level activity count}

\item{wear}{a data frame of similar strucutre to previous one containing
wear/nonwear flag}

\item{id}{id of the subject}

\item{day}{day number of the subject}

\item{h}{threshold determining sedentary and active, default is 100}
}
\value{
a plot
}
\description{
This function plots daily activity profile.
}
\examples{
data(act)
activity_profile(data=act,wear=flag,id=21062,day=1)


}
\keyword{activity_profile}

