% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actsum_avebyday.R
\name{actsum_avebyday}
\alias{actsum_avebyday}
\title{actsum_avebyday}
\usage{
actsum_avebyday(act_sum)
}
\arguments{
\item{act_sum}{dataframe containing subject-day summaries}
}
\value{
a dataframe ontaining subjects'summaries
}
\description{
This function average summaries across days for each subject
}
\examples{
data(act)
actsum_avebyday(data = act_sum)



}
\keyword{actsum_avebyday}

