% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validay.R
\name{validay}
\alias{validay}
\title{validay}
\usage{
validay(data, wear, hour = 10)
}
\arguments{
\item{data}{a data frame where each row is one subject day and
column 1 and 2 are id and day, and fowllowing colums
are minute level activity count}

\item{wear}{a data frame of similar strucutre to previous one containing
wear/nonwear flag}

\item{hour}{daily wearing time, default hour=10}
}
\value{
data frame of subject-day data
}
\description{
This function keeps subject-days that are valid
}
\examples{
data(act)
validay(data=act,wear=flag,hour=10)


}
\keyword{valid}

