% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNW.R
\name{WNW}
\alias{WNW}
\title{WNW}
\usage{
WNW(x, window = 90, tolerance = 2, cpmmax = 99)
}
\arguments{
\item{x}{a vector of activity counts}

\item{window}{window interval scanning, default is 90}

\item{tolerance}{number of maximum allowance in the window, default is 2}

\item{cpmmax}{tolerated maximum activity counts, default is 99}
}
\value{
a vector contating same length as x and its elements are either 1 or 0
}
\description{
This function generates wear/nonwera flags
}
\examples{
data(onedaydata)
WNW(x = xi,window = 90, tolerance = 2, cpmmax = 99)

}
\keyword{WNW}

