% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNW_data.R
\name{WNW_data}
\alias{WNW_data}
\title{WNW_data}
\usage{
WNW_data(data, window = 90, tolerance = 2, cpmmax = 99)
}
\arguments{
\item{data}{dataframe containing subject-day level activity count}

\item{window}{window interval scanning, default is 90}

\item{tolerance}{number of maximum allowance in the window, default is 2}

\item{cpmmax}{tolerated maximum activity counts, default is 99}
}
\value{
a vector contating same length as x and its elements are either 1 or 0
}
\description{
This function generates wear/nonwera flags
}
\examples{
data(act)
WNW_data(data = act, window = 90, tolerance = 2, cpmmax = 99)

}
\keyword{WNW_data}

